etl-service/requirements.txt
pandas
transformers
torch
google-generativeai
neo4j
streamlit